Atendendentes (sem o upload de foto)

+++ backend/src/database/migrations/20200412162121_create_atendente.js
@@ -1,7 +1,7 @@
 exports.up = knex => {
   return knex.schema.createTable('atendentes', table => {
     table.increments().primary();
-    table.integer('cpf').notNullable();
+    table.bigInteger('cpf').notNullable();
     table.string('nome').notNullable();
     table.string('email').notNullable();
     table.string('sexo').notNullable();
